+++ public/Arrays/MultiplePointer.js
