+++ public/Trees/index.js
@@ -1,0 +1,6 @@
+import './Search'
+import './TreeProblems'
+
+
+
+
