+++ public/index.html
@@ -17,27 +17,13 @@
     <br />
     MergeSort:
     <span id="MergeSort"></span>
-    <script src="DynamicProgramming.js"></script>
-    <!-- <script src="ArrayProblems.js"></script> -->
-    <!-- <script src="Graph.js"></script> -->
-    <!-- <script src="TreeProblems.js"></script> -->
-    <!-- <script src="ValidAnagrams.js"></script>
-    <script src="MultiplePointer.js"></script>
-    <script src="Recursion.js"></script>
-    <script src="Search.js"></script>
+    <script src="./DP/DynamicProgramming/index.js"></script>
+    <script src="./Arrays/index.js"></script>
+    <script src="./Sorting/index.js"></script>
+    <script src="./DP/index.js"></script>
+    <script src="./Graphs/index.js"></script>
+    <script src="./Strings/index.js"></script>
+    <script src="./Trees/index.js"></script>
     
-    <script src="PriorityQueue.js"></script>
-    <script src="DijkstrasAlgo.js"></script>
-    <script src="DijkMinHeap.js"></script>
-
-   
-    <script src="LinkedListProblems.js"></script>
-    <script src="StackProblems.js"></script>
-
-    <script src="BubbleSort.js"></script>
-    <script src="InsertionSort.js"></script>
-    <script src="MergeSort.js"></script>
-    <script src="SelectionSort.js"></script> -->
-    <!-- <script src="StringProblems.js"></script> -->
   </body>
 </html>
